compiscript/requirements.txt
antlr4-python3-runtime==4.13.0
antlr4-python3-runtime==4.13.0
Flask==2.2.2
Flask-Cors==3.0.10
werkzeug==2.2.2